requirements.txt
asgiref==3.5.2
boto3==1.24.84
botocore==1.27.84
Django==4.1
django-filter==22.1
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
jmespath==1.0.1
Pillow==9.2.0
PyJWT==2.4.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
